+++ obc/drivers/vn100/vn100_common.c
@@ -1,5 +1,4 @@
 #include "obc_errors.h"
-#include "obc_assert.h"
 #include "obc_board_config.h"
 #include "obc_logging.h"
 #include "obc_sci_io.h"
