+++ test/test_obc/unit/CMakeLists.txt
@@ -3,6 +3,9 @@ set(TEST_BINARY obc-firmware-tests)
 set(TEST_DEPENDENCIES
     ${CMAKE_SOURCE_DIR}/obc/sys/time/obc_time_utils.c
     ${CMAKE_SOURCE_DIR}/obc/drivers/arducam/image_processing.c
+    ${CMAKE_SOURCE_DIR}/obc/drivers/vn100/vn100_binary_parsing.c
+    ${CMAKE_SOURCE_DIR}/interfaces/obc_gs_interface/common/obc_gs_crc.c
+    ${CMAKE_SOURCE_DIR}/interfaces/data_pack_unpack/data_unpack_utils.c
 )
 
 set(TEST_MOCKS
@@ -13,6 +16,7 @@ set(TEST_SOURCES
     ${CMAKE_SOURCE_DIR}/test/test_obc/unit/main.cpp
     ${CMAKE_SOURCE_DIR}/test/test_obc/unit/test_obc_time_utils.cpp
     ${CMAKE_SOURCE_DIR}/test/test_obc/unit/test_image_processing.cpp
+    ${CMAKE_SOURCE_DIR}/test/test_obc/unit/test_vn100_unpack.cpp
 )
 
 set(TEST_SOURCES ${TEST_SOURCES} ${TEST_DEPENDENCIES} ${TEST_MOCKS})
@@ -25,6 +29,9 @@ target_include_directories(${TEST_BINARY}
     ${CMAKE_SOURCE_DIR}/obc/sys/time
     ${CMAKE_SOURCE_DIR}/obc/drivers/ds3232
     ${CMAKE_SOURCE_DIR}/obc/drivers/arducam
+    ${CMAKE_SOURCE_DIR}/obc/drivers/vn100
+    ${CMAKE_SOURCE_DIR}/interfaces/obc_gs_interface/common
+    ${CMAKE_SOURCE_DIR}/interfaces/data_pack_unpack
 )
 
 target_link_libraries(${TEST_BINARY}
