+++ test/test_obc/unit/test_vn100_unpack.cpp
@@ -1,7 +1,5 @@
-#include "vn100_common.h"
-#include "obc_sci_io.h"
+#include "vn100_binary_parsing.h"
 #include "obc_errors.h"
-#include "obc_gs_crc.h"
 
 #include <stdint.h>
 #include <stdlib.h>
@@ -9,11 +7,15 @@
 #include <gtest/gtest.h>
 
 TEST(TestVn100PackUnpack, ValidVn100PackUnpack) {
-  obc_error_code_t err;
+  const uint8_t header[] = {0xFA, 0x01, 0x28, 0x05};
+  const uint8_t payload[] = {0xA8, 0x02, 0xC3, 0x3F, 0xE7, 0x89, 0xB0, 0x42, 0x3D, 0xA9, 0x6A, 0xC1, 0xD0, 0x83,
+                             0x54, 0x3A, 0x92, 0x3E, 0x0B, 0xBB, 0x1C, 0x25, 0x30, 0xBA, 0x04, 0x7D, 0x1C, 0x41,
+                             0x3B, 0xB2, 0x83, 0x3D, 0x4E, 0xEA, 0x9C, 0xBE, 0xF1, 0xBD, 0x93, 0xBE, 0xA4, 0x55,
+                             0x8C, 0xBD, 0x26, 0xC7, 0x77, 0x3E, 0xF4, 0x28, 0xE6, 0x41, 0x57, 0xCE, 0xC1, 0x42};
+  const uint8_t crc[] = {0xC0, 0xE9};
+  float payloadFloat[14];
 
-  const uint8_t header [] = {0xFA, 0x01, 0x0528};
-  const float payload [14] = {35.0300, 32.9274, 73.3720, 7.0980, 39.3274, 16.7880, 84.7893, 40.1305, 82.1947, 89.9776, 57.8206, 47.1948, 81.2268, 26.0413}; 
-  uint16_t crc = calculateCrc16Ccitt(&payload, sizeof(payload));
+  memcpy(payloadFloat, payload, sizeof(payload));
 
   // Calculate the total size of the data array
   size_t totalSize = sizeof(header) + sizeof(payload) + sizeof(crc);
@@ -26,28 +28,27 @@ TEST(TestVn100PackUnpack, ValidVn100PackUnpack) {
   memcpy(mockVn100Packet + sizeof(header), payload, sizeof(payload));
   memcpy(mockVn100Packet + sizeof(header) + sizeof(payload), &crc, sizeof(crc));
 
-  initVn100();
-
   vn100_binary_packet_t receivedPacket;
+  vn100_error_t vnError;
 
-  ASSERT_EQ(readBinaryOutputs(&receivedPacket), OBC_ERR_CODE_SUCCESS); 
-  EXPECT_EQ(receivedPacket.yaw, payload[0]);
-  EXPECT_EQ(receivedPacket.pitch, payload[1]);
-  EXPECT_EQ(receivedPacket.roll, payload[2]);
+  ASSERT_EQ(parsePacket(mockVn100Packet, &receivedPacket, &vnError), OBC_ERR_CODE_SUCCESS);
 
-  EXPECT_EQ(receivedPacket.gyroX, payload[3]);
-  EXPECT_EQ(receivedPacket.gyroY, payload[4]);
-  EXPECT_EQ(receivedPacket.gyroZ, payload[5]);
+  EXPECT_EQ(receivedPacket.yaw, payloadFloat[0]);
+  EXPECT_EQ(receivedPacket.pitch, payloadFloat[1]);
+  EXPECT_EQ(receivedPacket.roll, payloadFloat[2]);
 
-  EXPECT_EQ(receivedPacket.accelX, payload[6]);
-  EXPECT_EQ(receivedPacket.accelY, payload[7]);
-  EXPECT_EQ(receivedPacket.accelZ, payload[8]);
+  EXPECT_EQ(receivedPacket.gyroX, payloadFloat[3]);
+  EXPECT_EQ(receivedPacket.gyroY, payloadFloat[4]);
+  EXPECT_EQ(receivedPacket.gyroZ, payloadFloat[5]);
 
-  EXPECT_EQ(receivedPacket.magX, payload[9]);
-  EXPECT_EQ(receivedPacket.magY, payload[10]);
-  EXPECT_EQ(receivedPacket.magX, payload[11]);
+  EXPECT_EQ(receivedPacket.accelX, payloadFloat[6]);
+  EXPECT_EQ(receivedPacket.accelY, payloadFloat[7]);
+  EXPECT_EQ(receivedPacket.accelZ, payloadFloat[8]);
 
-  EXPECT_EQ(receivedPacket.temp, payload[12]);
-  EXPECT_EQ(receivedPacket.pres, payload[13]);
+  EXPECT_EQ(receivedPacket.magX, payloadFloat[9]);
+  EXPECT_EQ(receivedPacket.magY, payloadFloat[10]);
+  EXPECT_EQ(receivedPacket.magZ, payloadFloat[11]);
 
+  EXPECT_EQ(receivedPacket.temp, payloadFloat[12]);
+  EXPECT_EQ(receivedPacket.pres, payloadFloat[13]);
 }
